requirements.txt
Flask
qrcode
Pillow
Werkzeug